script du choix du jeu en fonction du jeu sélectionner pour éviter les erreurs sur les autres jeux
