+++ assets/js/games/scriptAjaxScore.js
@@ -1,19 +1,5 @@
 export function saveScore(score, nameGame) {
-
     const requete = new XMLHttpRequest();
-    requete.open('GET', '/games/saveScore?score='+score+'&name_game='+nameGame, true)
-    requete.onreadystatechange = function() {
-        if(requete.readyState === 4 && requete.status === 200) {
-            const response = JSON.parse(requete.responseText);
-            if(response.success) {
-                console.log("Evenement symfony marche bien");
-            }
-        }
-    }
-
+    requete.open('GET', '/saveScore?score='+score+'&name_game='+nameGame, true)
     requete.send();
-
 }
-
-
-
