+++ assets/js/games/selectGame.js
@@ -1,0 +1,12 @@
+if(document.body.contains(document.querySelector('canvas'))) {
+    const canvas = document.querySelector('canvas')
+    switch(canvas.getAttribute("id")) {
+    case "snake":
+        require('./snake/snake');
+        break
+    case "planet-defense":
+        require('./planet_defense/planet_defense');
+        break
+    }
+}
+
