+++ templates/games/show.html.twig
@@ -52,15 +52,15 @@
                 </table>
             </div>
             <div class="game-podium">
-                {% if bestUsers is not empty %}
+                {% if plays is not empty %}
                     <div class="game-podium-name">
-                            {% for key, bestUser in bestUsers %}
+                            {% for key, bestUser in plays|slice(0, 3) %}
                                 {% if key == 0 %}
-                                    <h3 class="first-place">{{ bestUser }}</h3>
+                                    <h3 class="first-place">{{ bestUser.user.username }}</h3>
                                 {% elseif key == 1 %}
-                                    <h4 class="second-place">{{ bestUser }}</h4>
+                                    <h4 class="second-place">{{ bestUser.user.username }}</h4>
                                 {% elseif key == 2 %}
-                                    <h4 class="third-place">{{ bestUser }}</h4>
+                                    <h4 class="third-place">{{ bestUser.user.username }}</h4>
                                 {% endif %}
                             {% endfor %}
                     </div>
